Rotation encryption can encrypt messages without white spaces
Substitution encryption key is being worked on, the normal alphabet string is correctly defined and may be later defined in main using a separate function and then be passed onto this function. 

The encryption key still needs to work. The hardest part seems to be assigning a new value that is different from the last or cannot be the same.