Serious alterations made.
Final output now contains the original message next to the ciphered message as well as the key of the cipher compared to a regular alphabet (no scale given for rotation ciphers [yet?])

Added a nearly 'functional' rotation cipher decryption function when the cipher is given and currently only fails when white space exists in the entered message that needs to be decrypted and currently skips over this element. Which is the main bug to currently fix.
There are also some other minor bugs to look at.

Other decrypting functions have been theorised and will be ready to coded soon, especially the substitution ciphers and even the rotation cipher without the key.