Fixed the issue with the first rotation decryption when the cipher is given and the white spaces weren't recorded
Added a functional substitution encrypting function when the cipher and text is given.

Occasional bugs occur with the string origin_mes which is meant to record the original message to inform the user later of their encryption (this also helps in checking work), as often an additional symbol is added after the recorded message - but not the edited one.
I tried to fix this using the stringcopy command from the string.h library, but it presents many bugs when I try to use it - hence I will try to get a better grasp of the function.