Many quality of life and program changes - no new cryptography functions added
The input and cipher key is now connected via file I/O, however the output still uses stdout currently.

Changed the Alphabet_f to startupf, acting as a startup function that acts outside of main, simply to clean up what was a messy main - considering how it is now plagued with file I/O (which may also be moved there soon).
Because of this function and file I alterations, the individual cryptography functions are now much shorter as they do not have to account for lower case letters in the input or asking the user for the key anymore.

It may not show changes as I tried to commit to the repository for the two txt files, however, it did not work out so well.